requirement.txt
livekit-agents
python-dotenv
deepgram
openai
silero
